requirements.txt
streamlit
python-dotenv
openai>=1.0.0
requests==2.32.3
fpdf2
openai
